Various enhancements
Changes:
- Reworked default colors to better match 2048's
- Added option to specify font for number tiles
- Tile number label will now auto-shrink text to fit
- Fixed warnings involving casting between NSInteger and C default types
- Log statements no longer outputted for non-debug builds

+++ NumberTileGame/NumberTileGame/Models/F3HGameModel.m
@@ -73,7 +73,7 @@
     NSInteger row = 0;
     BOOL shouldExit = NO;
     while (YES) {
-        row = arc4random_uniform(self.dimension);
+        row = arc4random_uniform((uint32_t)self.dimension);
         // Check if row has any empty spots in column
         for (NSInteger i=0; i<self.dimension; i++) {
             if ([self tileForIndexPath:[NSIndexPath indexPathForRow:row inSection:i]].empty) {
@@ -88,7 +88,7 @@
     NSInteger column = 0;
     shouldExit = NO;
     while (YES) {
-        column = arc4random_uniform(self.dimension);
+        column = arc4random_uniform((uint32_t)self.dimension);
         if ([self tileForIndexPath:[NSIndexPath indexPathForRow:row inSection:column]].empty) {
             shouldExit = YES;
             break;
